Fixed bugs, now runnable
