+++ src/main/java/no/uib/inf112/core/player/NonPlayer.java
@@ -1,9 +1,24 @@
 package no.uib.inf112.core.player;
 
+import no.uib.inf112.core.GameGraphics;
+import no.uib.inf112.core.map.cards.Card;
 import org.jetbrains.annotations.NotNull;
 
+
 public class NonPlayer extends Player {
+
+    private Card[] deck;
+
     public NonPlayer(int x, int y, @NotNull Direction direction) {
         super(x, y, direction);
+    }
+
+
+    @Override
+    public PlayerCard getNextCard(int id) {
+        if (deck == null) {
+            deck = GameGraphics.getRoboRally().getDeck().draw(MAX_PLAYER_CARDS);
+        }
+        return new PlayerCard(deck[id], this);
     }
 }
