+++ src/main/java/no/uib/inf112/core/player/UserPlayer.java
@@ -1,6 +1,8 @@
 package no.uib.inf112.core.player;
 
 import no.uib.inf112.core.GameGraphics;
+import no.uib.inf112.core.ui.CardContainer;
+import no.uib.inf112.core.ui.actors.cards.SlotType;
 import no.uib.inf112.core.ui.event.ControlPanelEventHandler;
 import no.uib.inf112.core.ui.event.ControlPanelEventListener;
 import no.uib.inf112.core.ui.event.events.PowerDownEvent;
@@ -8,8 +10,13 @@ import no.uib.inf112.desktop.Main;
 import org.jetbrains.annotations.NotNull;
 
 public class UserPlayer extends Player {
+
+    private CardContainer cards;
+
     public UserPlayer(int x, int y, @NotNull Direction direction) {
         super(x, y, direction);
+
+        cards = new CardContainer(this);
         if (!Main.HEADLESS) {
             ControlPanelEventHandler eventHandler = GameGraphics.getCPEventHandler();
             eventHandler.registerListener(PowerDownEvent.class, (ControlPanelEventListener<PowerDownEvent>) event -> {
@@ -40,4 +47,16 @@ public class UserPlayer extends Player {
         }
         GameGraphics.getRoboRally().getPlayerHandler().nextPlayer();
     }
+
+
+    @Override
+    public PlayerCard getNextCard(int id) {
+        return new PlayerCard(cards.getCard(SlotType.HAND, id), this);
+    }
+
+    @NotNull
+    public CardContainer getCards() {
+        return cards;
+    }
+
 }
